+++ tests/unit/utils/conftest.py
@@ -1,4 +1,5 @@
 from typing import Dict, List
+from unittest.mock import MagicMock
 
 import pytest
 from pyspark.sql import DataFrame, SparkSession
@@ -69,3 +70,19 @@ def utils__spark_dataframe_missing_batch_column(
         {"source_node": 0, "target_node": 0, "partition_col": "b", "final_group": 0},
     ]
     return spark_fixture.createDataFrame(data)
+
+
+@pytest.fixture(scope="function")
+def mock_spark_session_incompatible() -> MagicMock:
+    m = MagicMock(spec=SparkSession)
+
+    m.version = "3.3.0"
+    return m
+
+
+@pytest.fixture(scope="function")
+def mock_spark_session() -> MagicMock:
+    m = MagicMock(spec=SparkSession)
+
+    m.version = "3.5.0"
+    return m