+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Added
 
 * Add changelog and PR template
+* Add verification function to assert Spark version is compatible (>= 3.4.0)
+* Replace `final_group` column with `group` column
 * Add heatmap visualization module
 * Add example notebook demonstrating the heatmap module and updated README.md